move node configs to config/node
